Move instantiate_from_config into module
