+++ scripts/make_samples.py
@@ -3,7 +3,7 @@ import torch
 import numpy as np
 from omegaconf import OmegaConf
 from PIL import Image
-from main import instantiate_from_config, DataModuleFromConfig
+from taming.util import instantiate_from_config
 from torch.utils.data import DataLoader
 from torch.utils.data.dataloader import default_collate
 from tqdm import trange
