+++ scripts/sample_conditional.py
@@ -5,7 +5,7 @@ from omegaconf import OmegaConf
 import streamlit as st
 from streamlit import caching
 from PIL import Image
-from main import instantiate_from_config, DataModuleFromConfig
+from taming.util import instantiate_from_config
 from torch.utils.data import DataLoader
 from torch.utils.data.dataloader import default_collate
