+++ scripts/sample_fast.py
@@ -7,8 +7,8 @@ from PIL import Image
 from tqdm import tqdm, trange
 from einops import repeat
 
-from main import instantiate_from_config
 from taming.modules.transformer.mingpt import sample_with_past
+from taming.util import instantiate_from_config
 
 
 rescale = lambda x: (x + 1.) / 2.
