+++ taming/models/cond_transformer.py
@@ -3,8 +3,8 @@ import torch
 import torch.nn.functional as F
 import pytorch_lightning as pl
 
-from main import instantiate_from_config
 from taming.modules.util import SOSProvider
+from taming.util import instantiate_from_config
 
 
 def disabled_train(self, mode=True):
