+++ taming/util.py
@@ -1,3 +1,4 @@
+import importlib
 import os, hashlib
 import requests
 from tqdm import tqdm
@@ -142,6 +143,20 @@ def retrieve(
         return list_or_dict, success
 
 
+def get_obj_from_str(string, reload=False):
+    module, cls = string.rsplit(".", 1)
+    if reload:
+        module_imp = importlib.import_module(module)
+        importlib.reload(module_imp)
+    return getattr(importlib.import_module(module, package=None), cls)
+
+
+def instantiate_from_config(config):
+    if not "target" in config:
+        raise KeyError("Expected key `target` to instantiate.")
+    return get_obj_from_str(config["target"])(**config.get("params", dict()))
+
+
 if __name__ == "__main__":
     config = {"keya": "a",
               "keyb": "b",
@@ -154,4 +169,5 @@ if __name__ == "__main__":
     config = OmegaConf.create(config)
     print(config)
     retrieve(config, "keya")
+
